- second code patch from Brent Johnson (Nonomori), aimed at source compatibility with VS.NET 2005
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@520 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Tactical/Soldier Profile.cpp
@@ -808,10 +808,10 @@ UINT16 CalcCompetence( MERCPROFILESTRUCT * pProfile )
 	uiStats = ((2 * pProfile->bLifeMax) + pProfile->bStrength + pProfile->bAgility + pProfile->bDexterity + ((pProfile->bLeadership + pProfile->bWisdom) / 2)) / 3;
 
 	// marksmanship is very important, count it double
-	uiSkills = (UINT32) ((2   * (pow(pProfile->bMarksmanship, 3) / 10000)) +
-												1.5 *	(pow(pProfile->bMedical, 3) / 10000) +
-															(pow(pProfile->bMechanical, 3) / 10000) +
-															(pow(pProfile->bExplosive, 3) / 10000));
+	uiSkills = (UINT32) ((2   * (pow((double)pProfile->bMarksmanship, 3) / 10000)) +
+												1.5 *	(pow((double)pProfile->bMedical, 3) / 10000) +
+															(pow((double)pProfile->bMechanical, 3) / 10000) +
+															(pow((double)pProfile->bExplosive, 3) / 10000));
 
 	// action points
 	uiActionPoints = 5 + (((10 * pProfile->bExpLevel +
